pyproject.toml
[project]
name = "learn"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cryptography>=46.0.5",
    "pycryptodome>=3.23.0",
]
